feature: CommentCard
